// sdk/attest-ts/src/proto/layr_tpm_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file layr_tpm.proto (package layr_tpm, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file layr_tpm.proto.
 */
export const file_layr_tpm: GenFile = /*@__PURE__*/
  fileDesc("Cg5sYXlyX3RwbS5wcm90bxIIbGF5cl90cG0iRQoFUXVvdGUSDQoFcXVvdGUYASABKAwSDwoHcmF3X3NpZxgCIAEoDBIcCgRwY3JzGAMgASgLMg4ubGF5cl90cG0uUENScyJ9CgRQQ1JzEiAKBGhhc2gYASABKA4yEi5sYXlyX3RwbS5IYXNoQWxnbxImCgRwY3JzGAIgAygLMhgubGF5cl90cG0uUENScy5QY3JzRW50cnkaKwoJUGNyc0VudHJ5EgsKA2tleRgBIAEoDRINCgV2YWx1ZRgCIAEoDDoCOAEqMgoKT2JqZWN0VHlwZRISCg5PQkpFQ1RfSU5WQUxJRBAAEgcKA1JTQRABEgcKA0VDQxAjKkoKCEhhc2hBbGdvEhAKDEhBU0hfSU5WQUxJRBAAEggKBFNIQTEQBBIKCgZTSEEyNTYQCxIKCgZTSEEzODQQDBIKCgZTSEE1MTIQDWIGcHJvdG8z");

/**
 * @generated from message layr_tpm.Quote
 */
export type Quote = Message<"layr_tpm.Quote"> & {
  /**
   * @generated from field: bytes quote = 1;
   */
  quote: Uint8Array;

  /**
   * @generated from field: bytes raw_sig = 2;
   */
  rawSig: Uint8Array;

  /**
   * @generated from field: layr_tpm.PCRs pcrs = 3;
   */
  pcrs?: PCRs;
};

/**
 * Describes the message layr_tpm.Quote.
 * Use `create(QuoteSchema)` to create a new message.
 */
export const QuoteSchema: GenMessage<Quote> = /*@__PURE__*/
  messageDesc(file_layr_tpm, 0);

/**
 * @generated from message layr_tpm.PCRs
 */
export type PCRs = Message<"layr_tpm.PCRs"> & {
  /**
   * @generated from field: layr_tpm.HashAlgo hash = 1;
   */
  hash: HashAlgo;

  /**
   * @generated from field: map<uint32, bytes> pcrs = 2;
   */
  pcrs: { [key: number]: Uint8Array };
};

/**
 * Describes the message layr_tpm.PCRs.
 * Use `create(PCRsSchema)` to create a new message.
 */
export const PCRsSchema: GenMessage<PCRs> = /*@__PURE__*/
  messageDesc(file_layr_tpm, 1);

/**
 * @generated from enum layr_tpm.ObjectType
 */
export enum ObjectType {
  /**
   * @generated from enum value: OBJECT_INVALID = 0;
   */
  OBJECT_INVALID = 0,

  /**
   * @generated from enum value: RSA = 1;
   */
  RSA = 1,

  /**
   * @generated from enum value: ECC = 35;
   */
  ECC = 35,
}

/**
 * Describes the enum layr_tpm.ObjectType.
 */
export const ObjectTypeSchema: GenEnum<ObjectType> = /*@__PURE__*/
  enumDesc(file_layr_tpm, 0);

/**
 * @generated from enum layr_tpm.HashAlgo
 */
export enum HashAlgo {
  /**
   * @generated from enum value: HASH_INVALID = 0;
   */
  HASH_INVALID = 0,

  /**
   * @generated from enum value: SHA1 = 4;
   */
  SHA1 = 4,

  /**
   * @generated from enum value: SHA256 = 11;
   */
  SHA256 = 11,

  /**
   * @generated from enum value: SHA384 = 12;
   */
  SHA384 = 12,

  /**
   * @generated from enum value: SHA512 = 13;
   */
  SHA512 = 13,
}

/**
 * Describes the enum layr_tpm.HashAlgo.
 */
export const HashAlgoSchema: GenEnum<HashAlgo> = /*@__PURE__*/
  enumDesc(file_layr_tpm, 1);